fix some more 64bit warnings/issues
git-svn-id: https://svn.code.sf.net/p/pinmame/code/trunk@3795 6f51dfdd-75df-4273-8f3b-818b1e6aa8bf

+++ src/input.c
@@ -599,7 +599,7 @@ int seq_cmp(InputSeq* a, InputSeq* b)
 	return 0;
 }
 
-void seq_name(InputSeq* code, char* buffer, unsigned max)
+void seq_name(InputSeq* code, char* buffer, size_t max)
 {
 	int j;
 	char* dest = buffer;
